+++ cmd/metrics/loader_static.go
@@ -12,7 +12,6 @@ import (
 	"slices"
 	"strings"
 
-	"github.com/Knetic/govaluate"
 	mapset "github.com/deckarep/golang-set/v2"
 )
 
@@ -25,7 +24,7 @@ func (l *StaticLoader) Load(metricDefinitionOverridePath string, eventDefinition
 	if err != nil {
 		return nil, nil, fmt.Errorf("failed to load event group definitions: %w", err)
 	}
-	configuredMetricDefinitions, err := configureMetrics(loadedMetricDefinitions, uncollectableEvents, GetEvaluatorFunctions(), metadata)
+	configuredMetricDefinitions, err := configureMetrics(loadedMetricDefinitions, uncollectableEvents, metadata)
 	if err != nil {
 		return nil, nil, fmt.Errorf("failed to configure metrics: %w", err)
 	}
@@ -78,247 +77,10 @@ func loadMetricDefinitions(metricDefinitionOverridePath string, selectedMetrics 
 	} else {
 		metrics = metricsInFile
 	}
-	return
-}
-
-// configureMetrics prepares metrics for use by the evaluator, by e.g., replacing
-// metric constants with known values and aligning metric variables to perf event
-// groups
-func configureMetrics(loadedMetrics []MetricDefinition, uncollectableEvents []string, evaluatorFunctions map[string]govaluate.ExpressionFunction, metadata Metadata) (metrics []MetricDefinition, err error) {
-	// get constants as strings
-	tscFreq := fmt.Sprintf("%f", float64(metadata.TSCFrequencyHz))
-	var tsc string
-	switch flagGranularity {
-	case granularitySystem:
-		tsc = fmt.Sprintf("%f", float64(metadata.TSC))
-	case granularitySocket:
-		tsc = fmt.Sprintf("%f", float64(metadata.TSC)/float64(metadata.SocketCount))
-	case granularityCPU:
-		tsc = fmt.Sprintf("%f", float64(metadata.TSC)/(float64(metadata.SocketCount*metadata.CoresPerSocket*metadata.ThreadsPerCore)))
-	default:
-		err = fmt.Errorf("unknown granularity: %s", flagGranularity)
-		return
+	// abbreviate event names in metrics to shorten the eventual perf stat command line
+	for _, metric := range metrics {
+		metric.Expression = abbreviateEventName(metric.Expression)
 	}
-	coresPerSocket := fmt.Sprintf("%f", float64(metadata.CoresPerSocket))
-	chasPerSocket := fmt.Sprintf("%f", float64(len(metadata.UncoreDeviceIDs["cha"])))
-	socketCount := fmt.Sprintf("%f", float64(metadata.SocketCount))
-	hyperThreadingOn := fmt.Sprintf("%t", metadata.ThreadsPerCore > 1)
-	threadsPerCore := fmt.Sprintf("%f", float64(metadata.ThreadsPerCore))
-	// load retire latency constants
-	var retireLatencies map[string]string
-	if retireLatencies, err = loadRetireLatencies(metadata); err != nil {
-		slog.Error("failed to load retire latencies", slog.String("error", err.Error()))
-		return
-	}
-	// configure each metric
-	reConstantInt := regexp.MustCompile(`\[(\d+)\]`)
-	for metricIdx := range loadedMetrics {
-		tmpMetric := loadedMetrics[metricIdx]
-		// abbreviate event names in metric expressions to match abbreviations used in uncollectableEvents
-		tmpMetric.Expression = abbreviateEventName(tmpMetric.Expression)
-		// skip metrics that use uncollectable events
-		foundUncollectable := false
-		for _, uncollectableEvent := range uncollectableEvents {
-			if strings.Contains(tmpMetric.Expression, uncollectableEvent) {
-				slog.Debug("removing metric that uses uncollectable event", slog.String("metric", tmpMetric.Name), slog.String("event", uncollectableEvent))
-				foundUncollectable = true
-				break
-			}
-		}
-		if foundUncollectable {
-			continue
-		}
-		// transform if/else to ?/:
-		var transformed string
-		if transformed, err = transformConditional(tmpMetric.Expression); err != nil {
-			return
-		}
-		// replace "> =" with ">=" and "< =" with "<="
-		transformed = strings.ReplaceAll(transformed, "> =", ">=")
-		transformed = strings.ReplaceAll(transformed, "< =", "<=")
-		if transformed != tmpMetric.Expression {
-			slog.Debug("transformed metric", slog.String("metric name", tmpMetric.Name), slog.String("transformed", transformed))
-			tmpMetric.Expression = transformed
-		}
-		// replace constants with their values
-		tmpMetric.Expression = strings.ReplaceAll(tmpMetric.Expression, "[SYSTEM_TSC_FREQ]", tscFreq)
-		tmpMetric.Expression = strings.ReplaceAll(tmpMetric.Expression, "[TSC]", tsc)
-		tmpMetric.Expression = strings.ReplaceAll(tmpMetric.Expression, "[CORES_PER_SOCKET]", coresPerSocket)
-		tmpMetric.Expression = strings.ReplaceAll(tmpMetric.Expression, "[CHAS_PER_SOCKET]", chasPerSocket)
-		tmpMetric.Expression = strings.ReplaceAll(tmpMetric.Expression, "[SOCKET_COUNT]", socketCount)
-		tmpMetric.Expression = strings.ReplaceAll(tmpMetric.Expression, "[HYPERTHREADING_ON]", hyperThreadingOn)
-		tmpMetric.Expression = strings.ReplaceAll(tmpMetric.Expression, "[CONST_THREAD_COUNT]", threadsPerCore)
-		tmpMetric.Expression = strings.ReplaceAll(tmpMetric.Expression, "[TXN]", fmt.Sprintf("%f", flagTransactionRate))
-		// replace retire latencies
-		for retireEvent, retireLatency := range retireLatencies {
-			// replace <event>:retire_latency with value
-			tmpMetric.Expression = strings.ReplaceAll(tmpMetric.Expression, fmt.Sprintf("[%s:retire_latency]", retireEvent), retireLatency)
-		}
-		// replace constant numbers masquerading as variables with their values, e.g., [20] -> 20
-		// there may be more than one with differing values in the expression, so use a regex to find them all
-		for {
-			// find the first match
-			found := reConstantInt.FindStringSubmatchIndex(tmpMetric.Expression)
-			if found == nil {
-				break // no more matches
-			}
-			// match[2] is the start of the number, match[3] is the end of the number
-			number := tmpMetric.Expression[found[2]:found[3]]
-			// replace the whole match with the number
-			tmpMetric.Expression = strings.ReplaceAll(tmpMetric.Expression, tmpMetric.Expression[found[0]:found[1]], number)
-		}
-		// get a list of the variables in the expression
-		tmpMetric.Variables = make(map[string]int)
-		expressionIdx := 0
-		for {
-			startVar := strings.IndexRune(tmpMetric.Expression[expressionIdx:], '[')
-			if startVar == -1 { // no more vars in this expression
-				break
-			}
-			endVar := strings.IndexRune(tmpMetric.Expression[expressionIdx:], ']')
-			if endVar == -1 {
-				err = fmt.Errorf("didn't find end of variable indicator (]) in expression: %s", tmpMetric.Expression[expressionIdx:])
-				return
-			}
-			// add the variable name to the map, set group index to -1 to indicate it has not yet been determined
-			tmpMetric.Variables[tmpMetric.Expression[expressionIdx:][startVar+1:endVar]] = -1
-			expressionIdx += endVar + 1
-		}
-		if tmpMetric.Evaluable, err = govaluate.NewEvaluableExpressionWithFunctions(tmpMetric.Expression, evaluatorFunctions); err != nil {
-			slog.Error("failed to create evaluable expression for metric", slog.String("error", err.Error()), slog.String("metric name", tmpMetric.Name), slog.String("metric expression", tmpMetric.Expression))
-			return
-		}
-		metrics = append(metrics, tmpMetric)
-	}
-	return
-}
-
-// transformConditional transforms if/else to ternary conditional (? :) so expression evaluator can handle it
-// simple:
-// from: <expression 1> if <condition> else <expression 2>
-// to:   <condition> ? <expression 1> : <expression 2>
-// less simple:
-// from: <expression 0> ((<expression 1>) if <condition> else (<expression 2>)) <expression 3>
-// to:   <expression 0> (<condition> ? (<expression 1>) : <expression 2) <expression 3>
-func transformConditional(origIn string) (out string, err error) {
-	numIfs := strings.Count(origIn, "if")
-	if numIfs == 0 {
-		out = origIn
-		return
-	}
-	in := origIn
-	for i := range numIfs {
-		if i > 0 {
-			in = out
-		}
-		var idxIf, idxElse, idxExpression1, idxExpression3 int
-		if idxIf = strings.Index(in, "if"); idxIf == -1 {
-			err = fmt.Errorf("didn't find expected if: %s", in)
-			return
-		}
-		if idxElse = strings.Index(in, "else"); idxElse == -1 {
-			err = fmt.Errorf("if without else in expression: %s", in)
-			return
-		}
-		// find the beginning of expression 1 (also end of expression 0)
-		var parens int
-		for i := idxIf - 1; i >= 0; i-- {
-			c := in[i]
-			switch c {
-			case ')':
-				parens += 1
-			case '(':
-				parens -= 1
-			default:
-				continue
-			}
-			if parens < 0 {
-				idxExpression1 = i + 1
-				break
-			}
-		}
-		// find the end of expression 2 (also beginning of expression 3)
-		parens = 0
-		for i, c := range in[idxElse+5:] {
-			switch c {
-			case '(':
-				parens += 1
-			case ')':
-				parens -= 1
-			default:
-				continue
-			}
-			if parens < 0 {
-				idxExpression3 = i + idxElse + 6
-				break
-			}
-		}
-		if idxExpression3 == 0 {
-			idxExpression3 = len(in)
-		}
-		expression0 := in[:idxExpression1]
-		expression1 := in[idxExpression1 : idxIf-1]
-		condition := in[idxIf+3 : idxElse-1]
-		expression2 := in[idxElse+5 : idxExpression3]
-		expression3 := in[idxExpression3:]
-		var space0, space3 string
-		if expression0 != "" {
-			space0 = " "
-		}
-		if expression3 != "" {
-			space3 = " "
-		}
-		out = fmt.Sprintf("%s%s%s ? %s : %s%s%s", expression0, space0, condition, expression1, expression2, space3, expression3)
-	}
-	return
-}
-
-type PlatformInfo struct {
-	ModelName      string `json:"Model name"`
-	CPUFamily      string `json:"CPU family"`
-	Model          string `json:"Model"`
-	ThreadsPerCore string `json:"Thread(s) per core"`
-	CoresPerSocket string `json:"Core(s) per socket"`
-	Sockets        string `json:"Socket(s)"`
-	Stepping       string `json:"Stepping"`
-	L3Cache        string `json:"L3 cache"`
-	NUMANodes      string `json:"NUMA node(s)"`
-	TMAVersion     string `json:"TMA version"`
-}
-
-type MetricStats struct {
-	Min  float64 `json:"MIN"`
-	Max  float64 `json:"MAX"`
-	Mean float64 `json:"MEAN"`
-}
-
-type RetireLatency struct {
-	Platform PlatformInfo           `json:"Platform"`
-	Data     map[string]MetricStats `json:"Data"`
-}
-
-func loadRetireLatencies(metadata Metadata) (retireLatencies map[string]string, err error) {
-	uarch := strings.ToLower(strings.Split(metadata.Microarchitecture, "_")[0])
-	uarch = strings.Split(uarch, " ")[0]
-	filename := fmt.Sprintf("%s_retire_latency.json", uarch)
-	var bytes []byte
-	if bytes, err = resources.ReadFile(filepath.Join("resources", "metrics", metadata.Architecture, metadata.Vendor, filename)); err != nil {
-		// not all architectures have retire latencies defined
-		err = nil
-		return
-	}
-	var retireLatency RetireLatency
-	if err = json.Unmarshal(bytes, &retireLatency); err != nil {
-		slog.Error("failed to unmarshal retire latencies", slog.String("error", err.Error()))
-		return
-	}
-	// create a map of retire latencies
-	retireLatencies = make(map[string]string)
-	for event, stats := range retireLatency.Data {
-		// use the mean value for the retire latency
-		retireLatencies[event] = fmt.Sprintf("%f", stats.Mean)
-	}
-	slog.Debug("loaded retire latencies", slog.Any("latencies", retireLatencies))
 	return
 }
 
@@ -395,38 +157,6 @@ func loadEventGroups(eventDefinitionOverridePath string, metadata Metadata) (gro
 		slog.Debug("Events not collectable on target", slog.String("events", uncollectable.String()))
 	}
 	return
-}
-
-// abbreviateEventName replaces long event names with abbreviations to reduce the length of the perf command.
-// focus is on uncore events because they are repeated for each uncore device
-func abbreviateEventName(event string) string {
-	// Abbreviations must be unique and in order. And, if replacing UNC_*, the abbreviation must begin with "UNC" because this is how we identify uncore events when collapsing them.
-	var abbreviations = [][]string{
-		{"UNC_CHA_TOR_INSERTS", "UNCCTI"},
-		{"UNC_CHA_TOR_OCCUPANCY", "UNCCTO"},
-		{"UNC_CHA_CLOCKTICKS", "UNCCCT"},
-		{"UNC_M_CAS_COUNT_SCH", "UNCMCC"},
-		{"IA_MISS_DRD_REMOTE", "IMDR"},
-		{"IA_MISS_DRD_LOCAL", "IMDL"},
-		{"IA_MISS_LLCPREFDATA", "IMLP"},
-		{"IA_MISS_LLCPREFRFO", "IMLR"},
-		{"IA_MISS_DRD_PREF_LOCAL", "IMDPL"},
-		{"IA_MISS_DRD_PREF_REMOTE", "IMDRP"},
-		{"IA_MISS_CRD_PREF", "IMCP"},
-		{"IA_MISS_RFO_PREF", "IMRP"},
-		{"IA_MISS_RFO", "IMRF"},
-		{"IA_MISS_CRD", "IMC"},
-		{"IA_MISS_DRD", "IMD"},
-		{"IO_PCIRDCUR", "IPCI"},
-		{"IO_ITOMCACHENEAR", "IITN"},
-		{"IO_ITOM", "IITO"},
-		{"IMD_OPT", "IMDO"},
-	}
-	// if an abbreviation key is found in the event, replace the matching portion of the event with the abbreviation
-	for _, abbr := range abbreviations {
-		event = strings.Replace(event, abbr[0], abbr[1], -1)
-	}
-	return event
 }
 
 // isCollectableEvent confirms if given event can be collected on the platform
@@ -588,4 +318,36 @@ func expandUncoreGroups(groups []GroupDefinition, metadata Metadata) (expandedGr
 		}
 	}
 	return
+}
+
+// abbreviateEventName replaces long event names with abbreviations to reduce the length of the perf command.
+// focus is on uncore events because they are repeated for each uncore device
+func abbreviateEventName(event string) string {
+	// Abbreviations must be unique and in order. And, if replacing UNC_*, the abbreviation must begin with "UNC" because this is how we identify uncore events when collapsing them.
+	var abbreviations = [][]string{
+		{"UNC_CHA_TOR_INSERTS", "UNCCTI"},
+		{"UNC_CHA_TOR_OCCUPANCY", "UNCCTO"},
+		{"UNC_CHA_CLOCKTICKS", "UNCCCT"},
+		{"UNC_M_CAS_COUNT_SCH", "UNCMCC"},
+		{"IA_MISS_DRD_REMOTE", "IMDR"},
+		{"IA_MISS_DRD_LOCAL", "IMDL"},
+		{"IA_MISS_LLCPREFDATA", "IMLP"},
+		{"IA_MISS_LLCPREFRFO", "IMLR"},
+		{"IA_MISS_DRD_PREF_LOCAL", "IMDPL"},
+		{"IA_MISS_DRD_PREF_REMOTE", "IMDRP"},
+		{"IA_MISS_CRD_PREF", "IMCP"},
+		{"IA_MISS_RFO_PREF", "IMRP"},
+		{"IA_MISS_RFO", "IMRF"},
+		{"IA_MISS_CRD", "IMC"},
+		{"IA_MISS_DRD", "IMD"},
+		{"IO_PCIRDCUR", "IPCI"},
+		{"IO_ITOMCACHENEAR", "IITN"},
+		{"IO_ITOM", "IITO"},
+		{"IMD_OPT", "IMDO"},
+	}
+	// if an abbreviation key is found in the event, replace the matching portion of the event with the abbreviation
+	for _, abbr := range abbreviations {
+		event = strings.Replace(event, abbr[0], abbr[1], -1)
+	}
+	return event
 }
