+++ cmd/metrics/metric.go
@@ -13,7 +13,6 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/Knetic/govaluate"
 	mapset "github.com/deckarep/golang-set/v2"
 )
 
@@ -82,46 +81,6 @@ func ProcessEvents(perfEvents [][]byte, eventGroupDefinitions []GroupDefinition,
 			slog.Debug("processed metric", slog.String("name", metricDef.Name), slog.String("expression", metricDef.Expression), slog.String("vars", strings.Join(prettyVars, ", ")))
 		}
 		metricFrames = append(metricFrames, metricFrame)
-	}
-	return
-}
-
-// GetEvaluatorFunctions defines functions that can be called in metric expressions
-func GetEvaluatorFunctions() (functions map[string]govaluate.ExpressionFunction) {
-	functions = make(map[string]govaluate.ExpressionFunction)
-	functions["max"] = func(args ...any) (any, error) {
-		var leftVal float64
-		var rightVal float64
-		switch t := args[0].(type) {
-		case int:
-			leftVal = float64(t)
-		case float64:
-			leftVal = t
-		}
-		switch t := args[1].(type) {
-		case int:
-			rightVal = float64(t)
-		case float64:
-			rightVal = t
-		}
-		return max(leftVal, rightVal), nil
-	}
-	functions["min"] = func(args ...any) (any, error) {
-		var leftVal float64
-		var rightVal float64
-		switch t := args[0].(type) {
-		case int:
-			leftVal = float64(t)
-		case float64:
-			leftVal = t
-		}
-		switch t := args[1].(type) {
-		case int:
-			rightVal = float64(t)
-		case float64:
-			rightVal = t
-		}
-		return min(leftVal, rightVal), nil
 	}
 	return
 }
